Removing code of hamburger menu and old course dashboard
