+++ Source/AccountViewController.swift
@@ -133,11 +133,9 @@ class AccountViewController: UIViewController, UITableViewDelegate, UITableViewD
                 launchEmailComposer()
             case .Logout:
                 OEXFileUtility.nukeUserPIIData()
-                if (environment.config.isTabLayoutEnabled) {
-                    dismiss(animated: true, completion: { [weak self] in self?.environment.router?.logout() })
-                } else {
-                    environment.router?.logout()
-                }
+                dismiss(animated: true, completion: { [weak self] in
+                    self?.environment.router?.logout()
+                })
             }
         }
         tableView.deselectRow(at: indexPath, animated: true)
