+++ Source/OEXRouter.h
@@ -17,7 +17,6 @@ NS_ASSUME_NONNULL_BEGIN
 @class OEXHelperVideoDownload;
 @class OEXVideoPathEntry;
 @class NetworkManager;
-@class RevealViewController;
 
 typedef NS_ENUM(NSUInteger, OEXSideNavigationState) {
     OEXSideNavigationStateHidden,
@@ -57,7 +56,6 @@ extern NSString* OEXSideNavigationChangedStateKey;
 - (void)showSignUpScreenFromController:(nullable UIViewController*)controller completion:(nullable void(^)(void))completion;
 
 #pragma mark Top Level
-- (void)showSidebar:(id)sender;
 - (void)showContentStackWithRootController:(UIViewController*)controller animated:(BOOL)animated;
 - (void)showMySettings;
 
@@ -73,7 +71,6 @@ extern NSString* OEXSideNavigationChangedStateKey;
 @interface OEXRouter (Private)
 
 @property (readonly, strong, nonatomic) RouterEnvironment* environment;
-@property (strong, nonatomic, nullable) RevealViewController* revealController;
 
 - (UINavigationController *) loginViewController;
 - (void)removeCurrentContentController;
@@ -86,7 +83,6 @@ extern NSString* OEXSideNavigationChangedStateKey;
 // UIViewController list for the currently navigation hierarchy
 - (NSArray*)t_navigationHierarchy;
 - (BOOL)t_showingLogin;
-- (BOOL)t_hasDrawerController;
 
 @end
