+++ Source/UserProfileViewController.swift
@@ -65,7 +65,6 @@ class UserProfileViewController: OfflineSupportViewController, UserProfilePresen
     }
     
     private func addBackBarButtonItem() {
-        if !environment.config.isTabLayoutEnabled { return }
         let backItem = UIBarButtonItem(image: Icon.ArrowLeft.imageWithFontSize(size: 40), style: .plain, target: nil, action: nil)
         backItem.oex_setAction {[weak self] in
             // Profile has different navbar color scheme that's why we need to revert nav bar color to original color while poping the controller
